Added CsvToXliff filter. Still need to write docs I promised around this, but the code is reasonably well documented. I haven't integrated this with the gui yet, so it's command line access only for now.
The filter has two modes of operation : bi-lingual and mono-lingual

For mono-lingual mode, we specify the list of translatable fields in a
properties file, and then the filter will put all of those fields into the
XLIFF file.

For bi-lingual operation, we specify the source field and the target field
and it will create an XLIFF file accordingly.

All non-translatable fields will be put into the skeleton file as formatting.
For now, we're not doing any segmentation of the contents of the CSV file,
this would be possible for mono-lingual operation, but is a bit more complex
for bi-lingual mode (as it would require sentence-alignment between the
source sentences and the target sentences)