Here is a comment of a user:
==============================================================
Could you check the segmentation for OLT analysis?

If there are following strings in a sgml file:
-------
Compartment bit 4 is the (inverse) bit for REL CNTRY1.
Compartment bit 4 being 0 means that the data is releasable to CNTRY1.
Compartment bit 5 is the (inverse) bit for REL CNTRY2.
Compartment bit 5 being 0 means that the data is releasable to CNTRY2.
Marking bit 11 is the (inverse) bit for REL CNTRY3.
Marking bit 11 being 0 means that the data is releasable to CNTRY3. Finally,
marking bit 13 is the NOFORN bit.
-------
I'd expect to have 6 segments in a .xlz file for them.
But the above strings are in ONE segment now.
===============================================================

    The segmenter is detecting the string 'CNTRY1.' as:

CNTRY: word
1.   : number

    I have fixed the code to make the segmentation this way:

CNTRY1: wordnumber
.     : segment_separator